Add files via upload
Designed a system to predict diseases based on the health data of a person using machine learning algorithms such as SVM, Naive Bayes Classifier and Random Forest Classifier.
In this projects, diseases are predicted based on the symptom or symptoms  a person is having. In this first all the three algorithms are implemented on the training data and its accuracy is calculated respectively.
Then all the three algorithms are implemented together on the test data to predict the final disease.